// Training/ecommerce/src/api/generated/model/iProduct.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * ECOMMERCE APP
 * The ECOMMERCE API description
 * OpenAPI spec version: 1.0
 */
import type { IProductProductStatus } from './iProductProductStatus';
import type { ICategory } from './iCategory';

export interface IProduct {
  id: number;
  productName: string;
  uploadDate: string;
  productDescription: string;
  price: number;
  sellerName: string;
  imageUrl: string;
  productStatus: IProductProductStatus;
  categories: ICategory[];
}